pyproject.toml
[project]
name = "notebookllm-production"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.7",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.26.7",
    "pypdf2>=3.0.1",
    "sentence-transformers>=5.2.0",
    "transformers>=4.57.6",
    "uvicorn>=0.40.0",
]
